cps: Make --with-library optionally work with a Content Server
This makes it possible to use the ebook conversion feature when an
instance of Calibre is running, by allowing the user to specify the
content server URL and credentials to perform updates with.
Previously this action would make calibredb error out and fail the
conversion job with an unknown error.